feat!: remove controller from builder
