+++ pubspec.yaml
@@ -3,7 +3,7 @@ description: Flutter Flows made easy! A Flutter package which simplifies flows w
 repository: https://github.com/felangel/flow_builder
 homepage: https://github.com/felangel/flow_builder
 
-version: 0.0.1-dev.1
+version: 0.0.1-dev.2
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
